cleaned up server side code with stripe routes
